finally refactored models.py into many, many files
