+++ network-api/networkapi/wagtailpages/pagemodels/customblocks/annotated_image_block.py
