+++ network-api/networkapi/wagtailpages/pagemodels/customblocks/image_text_block.py
