+++ network-api/networkapi/wagtailpages/pagemodels/customblocks/quote_block.py
